+++ common/src/main/java/com/github/tanhao1410/thesis/common/mapper/GroupDOMapper.java
@@ -44,7 +44,7 @@ public interface GroupDOMapper {
     /**
      * 插入行记录返回影响行数，使用getId()获取插入行记录的Id
      */
-    Integer insertSelectiveReturnPrimaryKey(GroupDO record);
+    Long insertSelectiveReturnPrimaryKey(GroupDO record);
 
     /**
      * 批量插入行数据返回插入行数
